2006-04-02  Alexis Sukrieh,,,  <sukria@backup-manager.org>
	* AUTHORS:
	  + added Brad Dixon about the S3 module.
	* VERSION:
	  + 0.7.3
	* Updated the user guide about the new purging system + minor changes.
	* lib/backup-methods.sh:
	  + Some messages updated.
	* lib/burning-methods.sh:
	  + Some messages updated.
	* lib/files.sh:
	  + Rewrite of the purging system, better handling, support the new policy
	  about master backups.
	* po/fr.po:
	  + Translation updates.
	* t/run-tests.sh:
	  + Better layout.




git-svn-id: svn://svn.backup-manager.org/backup-manager/trunk@384 2e458433-d701-0410-9826-f9b593394a3c

+++ t/t07-dar.sh
@@ -4,6 +4,7 @@ set -e
 
 # Each test script should include testlib.sh
 source testlib.sh
+
 # When the test is ready, set this to false for nice outputs.
 # if you want to see what happens, use those flags
 # verbose="true"
@@ -20,8 +21,12 @@ export BM_TARBALL_FILETYPE="dar"
 export BM_TARBALLINC_MASTERDATETYPE="weekly"
 export BM_TARBALLINC_MASTERDATEVALUE="1"
 
-# The test actions
+if [ ! -x $dar ]; then
+    info "cannot run test, need $dar"
+    exit 1
+fi
 
+# The test actions
 if [ -e $BM_ARCHIVE_ROOT ]; then
     rm -f $BM_ARCHIVE_ROOT/*
 fi
